requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-dotenv>=1.0.1
numpy>=1.24.0
pandas>=2.0.0
requests>=2.28.0
PyPDF2>=3.0.0
scikit-learn>=1.3.0
